Fixed wheel and pip handler to work with tensorflow and keras.
To work with tensorflow, we need to support the purelib format, which wasn't handled. This adds
another path in the import in py_library if the format (defined by the existance of a specific
folder) is found.

To work with keras, we need to inject the tensorflow dependency or the user will have to manually
spespecify the dependency on tensorflow in their own library to make it available in the local
namespace. This happens because tensorflow has 2 wheels: tensorflow and tensorflow-gpu. Keras can't
depend on either of them because it doesn't care if you're using GPUs or not. This occurs for every
python library that depends on tensorflow being available but can't express the "or" condition in
the requirements.

A new example is added to test both of these cases, showing they were broken before and showing they
work now.