Fixed wheel namespace for the case where collisions are desired
There are some family of packages, specially from Google, which rely on name collision in the
modules directories to work. For instance, `google-cloud` is actually an empty package that has
versioned dependencies to its "children" packages, all of which should be at `google/cloud` folder
inside pip. The same happens for many other Google packages, living in their own namespaces.

Since the default in this package is to have 1 namespace for each wheel, we end up with the issue
that only the first path that matches the module spec is considered. In the example provided,
`google.auth` is first in the search path (i.e. `sys.path`), so it's checked first. It indeed has a
`google` folder inside it with a package. Then Python tries to load `google.cloud`, which isn't
available there, so Python says it couldn't find the package. However, the original package was not
only in the search path, but it was actually the one listed in the dependencies!

This commit adds (along with some tech debt to be paid) support for a "dirty" namespace, where the
wheel and all of its dependencies are extracted in the same namespace, solving these issues. This
should be avoided in favor of cleanliness, but it's not always possible.