Add sitecustomize to get rid of dirty packages
Google packages start working if we evaluate *.pth files that
they provide.  Add a single sitecustomize.py file, and add it as a
dependency to every pip package.  sitecustomize.py (when found from
the PYTHONPATH) will be evaulated at the beginning, and we use it
to call site.addsitedir() to evaluate the *.pth files.