Fix BUILD file generation for dirty wheels
When extracting foo_dirty wheel repository, we pass all wheels as an
argument to piptool that transitively depend on foo, so that piptool
will extract all of them to the foo_dirty repository.  After that,
piptool generates a BUILD file from the *first* wheel, expecting
that to be foo.whl.  This was not happening, since we were passing
wheels in random order to piptool extract.