Add support for console entrypoints in wheels
Wheels can specify console entrypoints, which become executables people can call once it's
installed. It's very common to access functionality in this way directly, instead of as a lib
(example: tensorboard).

This commit parses the entrypoints described in the wheel and, for each one, creates a script that
will call the required function to simulate the entry point and adds an entry in the BUILD file for
that specific entrypoint. Note that we can't use the same entrypoint as the original script because
it might cause name conflicts with the folders, so we append "entrypoint_" to it to be explicit.