Use tree artifacts for extracted wheels to reduce the amount of symlinks
Previously, we would extract the wheel in the repository rule (during
loading/analysis phase) and define a py_library rule that globs all
extracted files into the "srcs" and "data" attributes.  This results in
each individual file being separately symlinked into the runfiles
directory, which creates LOTS of symlinks.

This change reduces the amount of symlinks by not extracting the wheel
in the repository rule, and instead instantiating a new "extract_wheel"
build rule that uses ctx.actions.declare_directory to declare a
directory artifact into which the wheel is extracted during the build
phase.

In the piptool code, this is implemented by splitting the old "piptool
extract" command into two: "piptool genbuild" that is ran in the loading
phase to generate the "extract_wheel" rule, and "piptool extract" that
is now invoked by the new "extract_wheel" rule in the build phase.
Patching also happens in the build phase now, so all patching commands
need to be propagated as attributes to the "extract_wheel" rule.